Basically got the core structure going
Able to load a list of images, manually set correlating points, and have
it produce warped output. Also added a separate script that uses
imagemagick and ffmpeg to produce a movie of the result.
Still a hack in there to force the bounding box - the calculation works
properly, but it needs to maintain the aspect ratio of the originals I
think